move MostPopularWords to lib
